operators/transcribe_speech_parakeet/app/requirements.txt
#torch==2.7.1
#numpy==2.1.1
nemo_toolkit[asr]
megatron-core
flask
